pyproject.toml
[project]
name = "reutov-max"
version = "0.1.0"
description = "Чат-бот мессенджера MAX для приёма заявок жителей города Реутов"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9",
    "httpx>=0.27",
    "openai>=1.55",
    "aiosqlite>=0.20",
    "pydantic>=2.7",
    "pydantic-settings>=2.4",
    "Pillow>=10.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
